+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## Version `1.0.2`
 
 - `storybook` Add `Select (custom)`
-- `ui` Add `Select (custom)`, `Badge`
+- `ui` Add `Select (custom)`, `Badge`, `VisuallyHidden`
+- `ui` Fix ARIA rules for `DialogContent`